Enhanced Article Relevance with User Tag Preferences (#24)
Incorporates user tag preferences into article relevance scoring.

This is similar to what we did with the user values. Tag Weights are calculated for the tags that the user cares about, using an exponential decay:
```csharp
var weight = 1.0 / Math.Pow(ExponentialDecayBase, Tag.Rank);
```

These tag weights are then used to calculate a  `tagAlignementScore` that is weighted/combined with the `valueAlignmentScore` and the `sentimentDelta` when calculating the overall article relevance.

## API changes
- Added a `ProfileService.GetUserProfileAsync()` method that is used by both the profile and article endpoints
- Removed `MinimumSentiment` and `SelectedValues` from the `RelevantArticlesRequest` *since these can be read directly from the user profile